requirements.txt
# Core
requests>=2.31.0
watchdog>=3.0.0
beautifulsoup4>=4.12.0

# Google APIs
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0

# Database
neo4j>=5.14.0

# AI / LangChain
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-core>=0.3.0

# Scheduling
APScheduler>=3.10.0

# Templating
Jinja2>=3.1.0

# Utilities
python-dotenv>=1.0.0
